initial commit: VolumeSlider
Credit's to Den on Stackoverflow, who did the Main Work on this Component.
This is Den's Profile: https://stackoverflow.com/users/2651521/den
This is Den's original post: https://stackoverflow.com/a/67734880/19068716

I did some beauty modifications that I think are awesome.
You can define the minimum and maximum value in the call paramters
You can also define the steps that the value changes when you move the button

You can add a text color
You can add a buttonColor or buttonGradientColor
You can add a activeBackgroundColor or a activeBackgroundGradientColor
You can add a passiveBackgroundColor or a passiveBackgroundGradientColor

This is my first package which I intend to use again myself, because I think
it is much better than the SwiftUI Slider Component.
There is however the UIKit Slider Component which seems to be better,
but I generally refuse to learn Legacy Code.